pyproject.toml
[build-system]
requires = ["scikit-build-core>=0.10", "pybind11"]
build-backend = "scikit_build_core.build"

[project]
name = "specfempp-core"
version = "0.3.0"
description = "Kokkos implementation of SPECFEM2D code"
readme = "README.md"
authors = [
  { name = "Rohit R Kakodkar", email = "rohit_kakodkar@princeton.edu" },
  { name = "Lucas Sawade", email = "lsawade@princeton.edu" },
  { name = "Congyue Cui", email = "ccui@princeton.edu" }
]
requires-python = ">=3.12,<3.13"

[tool.uv]
package = false

[dependency-groups]
all = [
  {include-group = "dev"},
  {include-group = "examples"},
  {include-group = "nightly-benchmarks"},
  {include-group = "scripts"}
]

dev = [
  {include-group = "style"},
  {include-group = "doc"},
]

style = [
  "pre-commit==2.19.0",
  "clang-format==19.1.7",
  "ruff==0.9.1",
]

doc = [
  "pre-commit==2.19.0",
  "clang-format==19.1.7",
  "ruff==0.9.1",
  "sphinx-rtd-theme==1.3.0",
  "doc8==0.11.2",
  "rstcheck==6.2.4",
  "restructuredtext-lint==1.4.0",
  "sphinx-copybutton==0.4.0",
  "furo==2024.8.6",
  "breathe==4.35.0",
  "sphinx-sitemap==2.6.0",
  "sphinx_design==0.6.1",
  "sphinxcontrib-video==0.4.1",
]

examples = [
  "snakemake==9.13.4",
  "obspy==1.4.1",
  "snakemake-executor-plugin-slurm==0.10.2",
  "imageio==2.36.1",
  "jinja2==3.1.6",
  "PyYAML==6.0.2",
  "numpy==2.2.4",
  "scipy==1.15.2",
  "h5py==3.13.0",
]

nightly-benchmarks = [
  {include-group = "examples"},
  "pandas==2.3.2",
]

scripts = [
  "gmsh==4.14.0",
  "matplotlib==3.10.3",
  "rtree==1.4.1",
]

[tool.scikit-build]
cmake.define.CMAKE_BUILD_TYPE = "Release"
cmake.define.SPECFEM_BUILD_TESTS = "ON"
cmake.define.SPECFEM_USE_SKBUILD = "ON"
cmake.define.SPECFEM_BINDING_PYTHON = "ON"
build-dir = "./build/python"
build.tool-args = ["-j4"]
